# openapi.yaml
basePath: /
consumes:
- application/json
definitions:
...
  Error:
    properties:
      error:
        description: Felmeddelande
        type: string
    type: object
...
info:
  contact:
    name: Ondfrukt
    url: https://github.com/ondfrukt/byggdelsuppdelning---demo
  description: REST API för Byggdelsuppdelning – ett PLM-system (Product Lifecycle
    Management) för hantering av byggdelar, objekttyper, relationer, dokument, klassificering
    och ändringshantering. Alla endpoints börjar med `/api/`.
  title: Byggdelsuppdelning API
  version: 1.0.0
paths:
...
  /api/view-config/list-view:
    get:
      responses:
        '200':
          description: Listvisningskonfiguration per objekttyp
          schema:
            additionalProperties:
              type: object
            type: object
        '500':
          description: Serverfel
          schema:
            $ref: '#/definitions/Error'
      summary: Get list view configuration for all object types
      tags:
      - View Configuration
    put:
      consumes:
      - application/json
      parameters:
      - in: body
        name: body
        required: true
        schema:
          additionalProperties:
            properties:
              column_order:
                items:
                  type: string
                type: array
              column_widths:
                type: object
              object_type_id:
                type: integer
              visible_columns:
                items:
                  type: object
                type: array
            type: object
          description: Nyckel = objekttypnamn, värde = konfiguration
          type: object
      responses:
        '200':
          description: Uppdaterade konfigurationer
          schema:
            properties:
              configs:
                items:
                  type: object
                type: array
              message:
                type: string
            type: object
        '400':
          description: Valideringsfel
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Serverfel
          schema:
            $ref: '#/definitions/Error'
      summary: Update list view configuration for object types
      tags:
      - View Configuration
  /api/view-config/list-view/{object_type_id}:
    get:
      parameters:
      - description: Objekttypens ID
        in: path
        name: object_type_id
        required: true
        type: integer
      responses:
        '200':
          description: Listvisningskonfiguration för typen
          schema:
            type: object
        '404':
          description: Objekttypen hittades inte
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Serverfel
          schema:
            $ref: '#/definitions/Error'
      summary: Get list view configuration for a specific object type
      tags:
      - View Configuration
  /api/view-config/tree-display:
    get:
      responses:
        '200':
          description: Trädvisningskonfiguration per objekttyp
          schema:
            additionalProperties:
              properties:
                available_fields:
                  items:
                    type: object
                  type: array
                object_type_id:
                  type: integer
                object_type_name:
                  type: string
                tree_view_name_field:
                  type: string
              type: object
            type: object
        '500':
          description: Serverfel
          schema:
            $ref: '#/definitions/Error'
      summary: Get tree display configuration for all object types
      tags:
      - View Configuration
    put:
      consumes:
      - application/json
      parameters:
      - in: body
        name: body
        required: true
        schema:
          additionalProperties:
            properties:
              object_type_id:
                type: integer
              tree_view_name_field:
                description: Fältnamn att visa i trädet (eller 'ID')
                type: string
            type: object
          description: Nyckel = objekttypnamn, värde = konfiguration
          type: object
      responses:
        '200':
          description: Uppdaterade konfigurationer
          schema:
            properties:
              configs:
                items:
                  type: object
                type: array
              message:
                type: string
            type: object
        '400':
          description: Valideringsfel
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Serverfel
          schema:
            $ref: '#/definitions/Error'
      summary: Update tree display configuration for object types
      tags:
      - View Configuration
produces:
- application/json
schemes:
- https
- http
securityDefinitions: {}
swagger: '2.0'
...